# api/0.2/swagger.yaml
openapi: 3.0.1
info:
  title: m2m
  description: API for M2M tests
  version: "0.2"
servers:
 -
  url: http://api.tsys.koeln:8280
security:
 -
  default: []
paths:
  /:
    get:
      parameters: []
      responses:
        "200":
          description: ok
      security:
       -
        default: []
      x-auth-type: Application & Application User
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
         - api_key
        optional: false
  /hello:
    get:
      parameters: []
      responses:
        "200":
          description: ok
      security:
       -
        default: []
      x-auth-type: Application & Application User
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
         - api_key
        optional: false
  /getWife:
    post:
      parameters: []
      responses:
        "200":
          description: ok
      security:
       -
        default: []
      x-auth-type: Application & Application User
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
         - api_key
        optional: false
  /getFamily:
    post:
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                person:
                  type: object
                  properties:
                    firstname:
                      type: string
                    lastname:
                      type: string
                    age:
                      type: number
                    isMarried:
                      type: boolean
                    gender:
                      type: string
        required: false
      responses:
        "200":
          description: ok
      security:
       -
        default: []
      x-auth-type: Application & Application User
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
         - api_key
        optional: false
components:
  securitySchemes:
    default:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ""
          scopes: {}
x-wso2-auth-header: Authorization
x-wso2-cors:
  corsConfigurationEnabled: false
  accessControlAllowOrigins:
   - '*'
  accessControlAllowCredentials: false
  accessControlAllowHeaders:
   - authorization
   - Access-Control-Allow-Origin
   - Content-Type
   - SOAPAction
   - apikey
   - testKey
  accessControlAllowMethods:
   - GET
   - PUT
   - POST
   - DELETE
   - PATCH
   - OPTIONS
x-wso2-production-endpoints:
  urls:
   - http://m2m.tsys.koeln
  type: http
x-wso2-sandbox-endpoints:
  urls:
   - http://m2m.tsys.koeln
  type: http
x-wso2-basePath: /m2m/0.2
x-wso2-transports:
 - http
 - https
x-wso2-response-cache:
  enabled: false
  cacheTimeoutInSeconds: 300